feat: Implement Firestore Query Builder and related components
- Added FirestoreValueCell for displaying Firestore values with support for various types.
- Introduced QueryBuilder component for building Firestore queries with filters, sorting, and limits.
- Created tests for QueryBuilder to ensure functionality and state management.
- Added types for Firestore queries and filters to improve type safety.
- Updated Firestore index to export new components and types.
- Refactored JSQuerySurface to utilize FirestoreDocumentBrowser for document previews.
- Introduced ChipList component for rendering lists of items with overflow handling.
- Added ExplorerTree component for hierarchical data representation with expandable rows.
- Implemented InspectorSection and DetailRow components for displaying metadata in a structured format.
- Updated package dependencies and lockfile to include new components.
