+++ packages/product-ui/package.json
@@ -24,6 +24,7 @@
     "@dnd-kit/core": "6.3.1",
     "@dnd-kit/sortable": "10.0.0",
     "@dnd-kit/utilities": "3.2.2",
+    "@firebase-desk/data-format": "workspace:*",
     "@firebase-desk/design-tokens": "workspace:*",
     "@firebase-desk/hotkeys": "workspace:*",
     "@firebase-desk/repo-contracts": "workspace:*",
