+++ packages/product-ui/src/features/firestore/index.ts
@@ -1,1 +1,3 @@
+export * from './FirestoreDocumentBrowser.tsx';
 export * from './FirestoreQuerySurface.tsx';
+export * from './types.ts';
