+++ packages/ui/src/chip-list/index.ts
@@ -1,0 +1,1 @@
+export * from './ChipList.tsx';
