+++ packages/ui/src/explorer-tree/index.ts
@@ -1,0 +1,1 @@
+export * from './ExplorerTree.tsx';
