+++ packages/ui/src/index.ts
@@ -1,14 +1,17 @@
 export * from './badge/index.ts';
 export * from './button/index.ts';
+export * from './chip-list/index.ts';
 export * from './cn.ts';
 export * from './context-menu/index.ts';
 export * from './data-table/index.ts';
 export * from './dialog/index.ts';
 export * from './dropdown-menu/index.ts';
 export * from './empty-state/index.ts';
+export * from './explorer-tree/index.ts';
 export * from './icon-button/index.ts';
 export * from './inline-alert/index.ts';
 export * from './input/index.ts';
+export * from './inspector/index.ts';
 export * from './JsonTreeView.tsx';
 export * from './panel/index.ts';
 export * from './resizable-panel-group/index.ts';
