+++ packages/ui/src/inspector/index.ts
@@ -1,0 +1,2 @@
+export * from './DetailRow.tsx';
+export * from './InspectorSection.tsx';
